==> 21_global.qmd <==
```{r set}
#| include: false
library(tidyverse)
```

# Global {.unnumbered}

For a first test, we used back global parameters calibrated for Paracou from @schmitt2024. A limit is the absence of calibration and tuning for mortality and crown radius allometry, which drive canopy packing, as well as the leaf-shedding parameters, which are important for TROLL4.0 dynamics. Moreover, as we will use texture only soil information from METRADICA, we used pedotransfer functions from Brooks & Corey instead of Van Genuchten Mualem.

```{r pars}
rcontroll::generate_parameters(
  CR_a = 1.80,
  CR_b = 0.3860,
  m = 0.035,
  m1 = 0.035,
  pheno_a0 = 0.2,
  pheno_b0 = 0.015,
  pheno_delta = 0.1,
  WATER_RETENTION_CURVE = 0
) %>%
  filter(!(param %in% c(
    "nbiter", "HEIGHT", "length_dcell", "NV", "NH", "nbout", "nbspp",
    "DBH0", "H0", "CR_min", "CD0", "_SOIL_LAYER_WEIGHT",
    "_WATER_RETENTION_CURVE", "_LL_parameterization", "_LA_regulation",
    "_sapwood", "_seedsadditional", "_NONRANDOM", "Rseed",
    "_GPPcrown", "_BASICTREEFALL", "_SEEDTRADEOFF",
    "_CROWN_MM", "_OUTPUT_extended", "_OUTPUT_inventory",
    "extent_visual"
  ))) %>%
  knitr::kable()
```
