==> 22_species.qmd <==
```{r set}
#| include: false
library(tidyverse)
library(ggfortify)
```

# Species {.unnumbered}

For a first test, species parameters are taken from @schmitt2024. A limit is the reduced species pool and functional space available compared to the whole French Guiana. This could be overcome using more information from @vleminckx2021 or Gentry inventories, as well as exploring back the approaches of Guillaume Salzet's PhD (pers. com.).

```{r species}
#| message: false
#| warning: false
#| fig-cap: "Caption."
data <- read_tsv("data/raw_data/Paracou_species_troll.tsv") %>% 
  select(-s_seedmass, -s_regionalfreq) %>% 
  rename_all(~ gsub("s_", "", .))
autoplot(
  princomp(select(data, -name), cor = TRUE),
  col = "lightgrey",
  loadings.label.size = 6,
  loadings.label.colour = "red", loadings.label.vjust = 1.1,
  loadings.label.repel = TRUE,
  loadings = TRUE, loadings.label = TRUE, loadings.colour = "red"
) +
  coord_equal() +
  geom_hline(aes(yintercept = 0), col = "black", linetype = "dotted") +
  geom_vline(aes(xintercept = 0), col = "black", linetype = "dotted") +
  theme_bw() +
  theme(legend.position = "none")
```
