==> 24_climate.qmd <==
```{r set}
#| include: false
library(tidyverse)
library(sf)
library(terra)
```

# Climate {.unnumbered}

ERA5-Land reanalysis [@muñoz-sabater2021] have a grid matching the 10-km grid we used. They show limited monthly variation in the test area but still a strong coast to inland gradient as well as a slight east-west gradient. Due to resolution issue we have to remove 4 grid cells that does not match ERA5-Land from our simulations, bringing them to 86.

```{r climate_monthly}
#| eval: false
climate <- vroom::vroom("data/raw_data/era/era.tsv") %>% 
  group_by(lon, lat, date = lubridate::floor_date(time, "month")) %>% 
  summarise(
    snet = mean(snet),
    tas = mean(tas),
    vpd = mean(vpd),
    ws = mean(ws),
    pr = sum(pr)
  )
write_tsv(climate, "data/derived_data/climate_monthly.tsv")
```

```{r monthly_var}
#| message: false
#| warning: false
#| fig-cap: "Caption."
climate <- read_tsv("data/derived_data/climate_monthly.tsv")
climate %>% 
  gather(variable, value, -lat, -lon, -date) %>% 
  ggplot(aes(date, value)) +
  geom_line(aes(group = paste(lon, lat)), alpha = .1) +
  facet_wrap(~variable, scales = "free_y", nrow = 5) +
  theme_bw() +
  xlab("") +
  ylab("")
```

```{r sp_var}
#| message: false
#| warning: false
#| fig-cap: "Caption."
cells <- read_tsv("data/derived_data/grid_era.tsv") %>% 
  rename(lon = X, lat = Y) %>% 
  mutate(value = 1)
climate %>% 
  filter(year(date) == 2024) %>% 
  group_by(lon, lat) %>% 
  summarise(
    snet = mean(snet),
    tas = mean(tas),
    vpd = mean(vpd),
    ws = mean(ws),
    pr = sum(pr)
  ) %>% 
  mutate(across(c("snet", "tas", "vpd", "ws", "pr"), scale)) %>% 
  gather(variable, value, -lon, -lat) %>% 
  ggplot(aes(lon, lat, fill = value)) +
  geom_raster() +
  geom_point(data = cells) +
  theme_bw() +
  facet_wrap(~ variable) +
  scale_fill_viridis_c(na.value = "transparent", guide = "none") +
  coord_equal() +
  xlab("") + ylab("")
```

```{r grid_era}
#| eval: false
non_empty <- extract(clim_r[[1]], vect(grid)) %>% 
  mutate(na = is.na(AGB_16122015)) %>% 
  group_by(ID) %>% 
  summarise(na = sum(na)/n()) %>% 
  filter(na <= 0.5)
clim_r <- climate %>% 
  filter(year(date) == 2024) %>% 
  group_by(lon, lat) %>% 
  summarise(
    snet = mean(snet),
    tas = mean(tas),
    vpd = mean(vpd),
    ws = mean(ws),
    pr = sum(pr)
  ) %>% 
  rast()
cells <- read_tsv("data/derived_data/grid.tsv") %>% 
  st_as_sf(coords = c("X", "Y"), crs = 4326)
non_empty <- extract(clim_r[[1]], cells) %>% 
  filter(!is.na(snet))
read_tsv("data/derived_data/grid.tsv")[non_empty$ID,] %>% 
  write_tsv("data/derived_data/grid_era.tsv")
coords_era <- extract(clim_r[[1]], cells, xy = TRUE) %>% 
  select(-ID, -snet) %>% 
  rename(lon = x, lat = y)
read_tsv("data/derived_data/grid.tsv") %>% 
  bind_cols(coords_era) %>% 
  write_tsv("data/derived_data/correspondence_era")
```
