==> 32_guyafor.qmd <==
```{r set}
#| include: false
library(tidyverse)
```

# Guyafor {.unnumbered}

Simulated forest dynamics and composition will be evaluated against several study areas with several censuses from the Guyafor network (<https://paracou.cirad.fr/website/experimental-design/guyafor-network>, Géraldine Derroire pers. com.). For the moment we will focus on Aboveground Carbon Stock (ACS) dynamics per plot per forest in time and its resulting variation across years $\Delta ACS$. However, based on the cleaning and taxonomy imputation scripts for inventories provided by Géraldine, we could further explore the distribution of diameters, basal area, tree growth and mortality, as well as the neutral and functional composition. The ACS data included values for each census year, as well as for forest, plot and 0.25-ha quadrat, and a simulation of taxonomic composition (due to the use of vernacular names). To be closer to the TROLL simulation, I summed the ACS back at the plot scale (approximately 4 ha) and computed the median and 5th and 95th quantiles across taxonomic composition for comparison with the simulation outputs.

```{r guyafor}
#| eval: false
coords <- read_tsv("data/raw_data/guyafor.tsv") %>% 
  select(Forest, Plot, Xutm, Yutm) %>% 
  group_by(Forest, Plot) %>% 
  summarise_all(mean, na.rm = TRUE) %>% 
  na.omit() %>% 
  rename_all(tolower)
qs::qread(
  file.path("data", "raw_data", "guyafor",
                               "Data_postStep4_ACS_perQuadrat_AllSites_100simx5iterACS.qs"
            )) %>% rename_all(tolower) %>% 
  group_by(sim, forest, plot, censusyear) %>% 
  summarise(acs = sum(acsperquadrat)/(n()*0.25)) %>% 
  group_by(forest, plot, censusyear) %>% 
  summarise(acs_q5 = quantile(acs, .05), 
            acs = median(acs),
            acs_q95 = quantile(acs, .95)) %>% 
  left_join(coords) %>% 
  write_tsv("data/derived_data/guyafor_acs.tsv")
qs::qread(
  file.path("data", "raw_data", "guyafor",
                               "Data_postStep4_ACS_perQuadrat_AllSites_100simx5iterACS.qs"
            )) %>% rename_all(tolower) %>% 
  group_by(sim, forest, plot, censusyear) %>% 
  summarise(acs = sum(acsperquadrat)/(n()*0.25)) %>% 
  group_by(sim, forest, plot) %>% 
  arrange(censusyear) %>% 
  mutate(delta_acs = (acs -  lag(acs))/(censusyear -  lag(censusyear))) %>% 
  na.omit() %>% 
  group_by(forest, plot, censusyear) %>% 
  summarise(delta_acs = median(delta_acs)) %>% 
  left_join(coords) %>% 
  write_tsv("data/derived_data/guyafor_delta_acs.tsv")
```

```{r acs}
#| message: false
#| warning: false
#| fig-cap: "Caption."
read_tsv("data/derived_data/guyafor_acs.tsv") %>% 
  ggplot(aes(censusyear, acs, group = plot)) +
  geom_ribbon(aes(ymin = acs_q5 , ymax = acs_q95), alpha=0.2) + 
  geom_line() +
  theme_bw() +
  scale_y_log10() +
  facet_wrap(~ forest) +
  xlab("") + ylab(expression("Aboveground Carbon Stock ["~t~ha^{-1}~"]"))
```

```{r delta_acs}
#| message: false
#| warning: false
#| fig-cap: "Caption."
read_tsv("data/derived_data/guyafor_delta_acs.tsv") %>% 
  ggplot(aes(delta_acs, forest)) +
  geom_boxplot() +
  theme_bw() +
  ylab("") +
  xlab(expression(Delta[ACS]~"["~t~ha^{-1}~yr^{-1}~"]"))
```

> GD: Kaw needs to be removed and Trinité and Trésor are spurious.
