==> 40_experiments.qmd <==
# Virtual experiments {.unnumbered}

For the test, the virtual experiment comprises a single simulation (with code built for replication) of a 4-hectare forest established over 600 years in representative soil and climate conditions within a 10 km cell. These forest simulations are then aggregated into summarised statistics of interest for evaluating fluxes, dynamics and spatial structure, as well as projecting the carbon balance. Future developments of the experiment could integrate future projections with climate change projections, beyond the limits specified below.

```{r scheme}
knitr::include_graphics("figures/fg_fig.png")
```

## Limits

The current limits of the experiment are the TROLL 4.0 representation of:

-   Water table depth
-   Soil depth
-   Species diversity, pool and functional trait diversity
-   Water stress response
-   Within cell variation
-   Temperature and water deficit response for future projection

## How to use

The virtual experiment relies on the reproducible and portable Snakemake workflow defined in the simulations folder, which is currently run on a high-performance computer. Further details can be found in the folder's readme file, but once the input data has been gathered (see simulations/data/README.md), usage is automated with a single bash script. However, I recommend carrying out a dry run of Snakemake first, as shown below:

``` bash
snakemake -np
sbatch job.sh
```
